refactor: Move model cache to separate module, create `generate_batch`
